requirements.txt
openai==0.27.0
transformers==4.30.2
torch==2.5.1
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
pydantic==1.10.8
spacy==3.5.0
nltk==3.8
langchain==0.0.85
llama-index==0.6.24
pdfplumber==0.9.0
celery==5.2.7
redis==4.5.5
diskcache==5.6.1
faiss-cpu==1.9.0
SpeechRecognition==3.14.0
pyttsx3==2.98
pyaudio==0.2.14